refactor(file-structure): rename folders and component to meet our naming convention
Closes #31
