+++ src/pages/Showcase/components/ShotCover/index.module.scss
