+++ src/pages/Showcase/components/ShotCover/index.tsx
