+++ src/pages/Showcase/components/ShotPreview/index.module.scss
