+++ src/pages/Showcase/components/ShotPreview/index.tsx
