+++ src/pages/Showcase/components/ShotTitle/index.module.scss
