+++ src/pages/Showcase/components/ShotTitle/index.tsx
