+++ src/pages/shots/MultiPlayerInteraction/components/Content/index.module.scss
