+++ src/pages/shots/MultiPlayerInteraction/components/Content/index.test.tsx
