+++ src/pages/shots/MultiPlayerInteraction/components/Content/index.tsx
