+++ src/pages/shots/MultiPlayerInteraction/components/JoystickIcon/index.module.scss
