+++ src/pages/shots/MultiPlayerInteraction/components/JoystickIcon/index.test.tsx
