+++ src/pages/shots/MultiPlayerInteraction/components/Player/index.module.scss
