+++ src/pages/shots/MultiPlayerInteraction/components/Player/index.test.tsx
