+++ src/pages/shots/MultiPlayerInteraction/components/Player/index.tsx
