+++ src/pages/shots/MultiPlayerInteraction/components/Players/index.module.scss
