+++ src/pages/shots/MultiPlayerInteraction/components/Players/index.test.tsx
