+++ src/pages/shots/MultiPlayerInteraction/components/Players/index.tsx
