+++ src/pages/shots/MultiPlayerInteraction/index.module.scss
