+++ src/pages/shots/index.ts
@@ -1,5 +1,5 @@
 import { Shot } from '../../common/interfaces/shot';
-import {shot as multiPlayerInteraction} from './multiplayer-interaction';
+import {shot as multiPlayerInteraction} from './MultiPlayerInteraction';
 
 export const allShots: Shot[] = [
   multiPlayerInteraction,
